Clock uses long instead of int for periods of time.
git-svn-id: https://sovereigncomputing.net/svn/sneer/trunk@3534 ee3b4d21-14f5-0310-a1fa-c5d8b2925ec4

+++ sneer/bricks/src/sneer/pulp/clock/Clock.java
@@ -8,14 +8,14 @@ public interface Clock extends Brick {
 	
 	long time();
 	
-	void sleepAtLeast(int millis);
+	void sleepAtLeast(long millis);
 
 	void wakeUpNoEarlierThan(long timeToWakeUp, Runnable runnable);
-	void wakeUpInAtLeast(int millisFromNow, Runnable runnable);
-	void wakeUpEvery(int minimumPeriodInMillis, Stepper stepper);
-	void wakeUpNowAndEvery(int minimumPeriodInMillis, Stepper stepper);
+	void wakeUpInAtLeast(long millisFromNow, Runnable runnable);
+	void wakeUpEvery(long minimumPeriodInMillis, Stepper stepper);
+	void wakeUpNowAndEvery(long minimumPeriodInMillis, Stepper stepper);
 
-	void advanceTime(int deltaMillis);
+	void advanceTime(long deltaMillis);
 	void advanceTimeTo(long absoluteTimeMillis);
 
 }
